SI-7261 BooleanSetting to Boolean conversion
Add an implicit for BooleanSetting to Boolean, since while we never
expect the Spanish Inquisition, we do expect a boolean in an if clause.

The first cut put the implicit at AbsScalaSettings, but the more
general case is at reflect/internal's MutableSettings.  There was no
particular desire to privilege mutability.