SI-7261 Applications of BooleanSetting to Boolean conversion
Most BooleanSettings are accessed in a conditional, but the verbiage
of settings.myflag.value gave rise frequent tmp vals.  These are
usually eliminated in favor of an import, hence if (myflag).

Convert BooleanSetting to BooleanFlag.  You'd think that would be trivial, if necessary.

There are a few minor s"$tweaks" for readability.

A few trailing case ex: that confuse vi modelines are mitigated.