SI-6289 Paulptest demonstrating javac errors
This is Paul's test demonstrating that Javac errors are
correctly transcribed in the test transcript.

A gratuitous Scala class is added to a later round to
show that the test halts after the first error.

The runner must supply absolute paths to javac so that
absolute paths are reported in errors and stripped away
by partest.

The check file is differentiated for Java 6 and 7,
and partest's runner will now post-process the
`diff log check` to strip the diff which does not apply.