SI-7300 single line comment in multi line comment
This issue was fixed in 3d5c675982 but didn't get any test cases,
which are added by this commit.

Before, a single line comment that occurred before a closing multi
line comment, like in `/*//*/`, was not treated as the beginning of
a nested comment, thus the shown example was parsed as a valid
comment.