SI-7335 Don't import Predef._ in Predef.scala
The check for this was foiled by the introduction of
DeprecatedPredef in c26a8db0. This didn't cause any
harm, but in any case this commit restores the intended
state of affairs.

The change is visible only through logging; no test case
is included.

    qbin/scalac -Ydebug -Ylog:all src/library/scala/Predef.scala -Ystop-after:typer 2>&1 | grep -i "Omitted import of Predef"
    [log namer] Omitted import of Predef._ for Predef.scala
    [log typer] Omitted import of Predef._ for Predef.scala

A redundant override of `firstDefinesClassOrObject` is also
removed.

Maybe we can remove the special case altogether?