Fix SI-6387 taking into account SI-7341.
In 4e10b2c83 the logic of expandName was changed to preserve
overriding symbols when expanding. This change broke the invariant
that the entire (field, getter, setter) triplet should be expanded
if any of its components is.

Fix this invariant by expanding all above symbols for `accessed` in
a later stage, if an initially expanded symbol is an accessor.