SI-7349 Partest supports test-interface
This commit updates Scalacheck to 1.10.1 and invokes it using the
test-interface API.  This obviates kludgy output checking for result
status.

The context class loader is set to a loader that the scaladoc scalacheck
tests can use, but those tests don't need to be scalacheck tests (as they
merely assert a predicate).  A useful refactor would reunite the source
submitted to scaladoc and the test making an assertion about the result,
but that must wait for GUTtest, which should run scaladoc the way
scalac is currently run in compilation rounds.